requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
dask==2.1.0
decorator==4.4.0
dialogflow==0.6.0
Flask==1.1.1
future==0.17.1
google-api-core==1.14.2
google-auth==1.6.3
googleapis-common-protos==1.6.0
mongoengine==0.18.2
confuse==1.0.0
gunicorn==19.4.5
h5py==2.9.0
html5lib==1.0.1
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mock==3.0.5
mpmath==1.1.0
multipledispatch==0.6.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
numba==0.44.1
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
opencv-python==4.1.0.25
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
ply==3.11
py==1.8.0
pycrypto==2.6.1
pycurl==7.43.0.3
pylint==2.3.1
pyramid==1.10.4
PySocks==1.7.0
python-dateutil==2.8.0
python-telegram-bot==12.0.0
PyYAML==5.1.1
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
SQLAlchemy==1.3.5
statsmodels==0.10.0
tables==3.5.2
tblib==1.4.0
urllib3==1.24.2
python-dotenv==0.10.3